Add horrible TODO for the TESR rendering.
At certain distances and angles, the Marker TE is no longer in the rendered chunks list and therefore the TESR part is not called, but my rendering should be displayed even when the TE should not be. Even settings getRenderingBoundingBox  to INFINITE did not work. Debugging showed that the getRenderingBoundBox was not called at certain angles.